new component, implement storage model for keycloak
 - sfdsAuth is now it's own component
 - added functionality to chunk the sfds authorizations for storage in Keycloak attributes
 - added functionality to delete an sfds auth and have it persist in Keycloak
