+++ frontend/src/components/UserUpdateRoles.vue
@@ -86,104 +86,7 @@
       </div>
       <span v-if="selectedClient.clientId.includes('SFDS')">
         <v-divider class="sub-permissions"></v-divider>
-
-        <v-data-table
-            :headers="sfdsTableHeaders"
-            :items="sfds_authorizations">
-          <template v-slot:top>
-            <v-toolbar
-                flat
-            >
-              <h2 class="sfds-authorizations-header">SFDS Authorizations</h2>
-              <v-spacer></v-spacer>
-              <v-dialog v-model="editDialog" max-width="840px">
-                <template v-slot:activator="{ on, attrs }">
-                  <v-btn color="primary" darkclass="mb-2" v-bind="attrs" v-on="on">
-                    New Authorization
-                  </v-btn>
-                </template>
-                <v-card>
-                  <v-card-title>
-                    <span class="headline">{{ dialogTitle }}</span>
-                  </v-card-title>
-                  <v-card-text>
-                    <v-container>
-                      <v-row class="right-gutters">
-                        <v-col class="col-4">
-                          <label>Mailbox</label>
-                          <v-autocomplete
-                              id="sfds-mailboxes"
-                              v-model="currentSfdsAuthorization.mailbox"
-                              outlined
-                              dense
-                              :items="sfdsMailboxes"
-                          ></v-autocomplete>
-                        </v-col>
-                        <v-col class="col-4">
-                          <label>Use</label>
-                          <v-autocomplete
-                              id="sfds-uses"
-                              v-model="currentSfdsAuthorization.uses"
-                              outlined
-                              dense
-                              multiple
-                              chips
-                              :items="sfdsUses"
-                          ></v-autocomplete>
-                        </v-col>
-                        <v-col class="col-4">
-                          <label>Permission</label>
-                          <v-select
-                              id="sfds-permissions"
-                              v-model="currentSfdsAuthorization.permission"
-                              outlined
-                              dense
-                              :items="sfdsPermissions"
-                          ></v-select>
-                        </v-col>
-                      </v-row>
-                    </v-container>
-                  </v-card-text>
-                  <v-card-actions>
-                    <v-btn class="primary" @click="saveSfdsAuthorization">
-                      Save
-                    </v-btn>
-                    <v-btn outlined class="primary--text" @click="close">
-                      Cancel
-                    </v-btn>
-                  </v-card-actions>
-                </v-card>
-              </v-dialog>
-              <v-dialog v-model="deleteDialog" max-width="650px" >
-                <v-card >
-                  <v-card-title class="headline">Are you sure you want to delete this authorization?</v-card-title>
-                  <v-card-text class="black--text">
-                    Use: {{ currentSfdsAuthorization.uses }} <br/>
-                    Mailbox: {{ currentSfdsAuthorization.mailbox }} <br/>
-                    Permission: {{ currentSfdsAuthorization.permission }}
-                  </v-card-text>
-                  <v-card-actions>
-                    <v-btn class="red white--text" @click="deleteItemConfirm">Delete</v-btn>
-                    <v-btn outlined class="primary--text" @click="closeDelete">Cancel</v-btn>
-                  </v-card-actions>
-                </v-card>
-              </v-dialog>
-            </v-toolbar>
-          </template>
-          <template v-slot:item.actions="{ item }">
-            <v-icon
-                small
-                class="mr-2"
-                @click="editItem(item)">
-              mdi-pencil
-            </v-icon>
-            <v-icon
-                small
-                @click="deleteItem(item)">
-              mdi-delete
-            </v-icon>
-          </template>
-        </v-data-table>
+        <user-update-sfds-auth></user-update-sfds-auth>
       </span>
     </div>
 
@@ -193,9 +96,11 @@
 <script>
 import ClientsRepository from "@/api/ClientsRepository";
 import UsersRepository from "@/api/UsersRepository";
+import UserUpdateSfdsAuth from "@/components/UserUpdateSfdsAuth";
 
 export default {
   name: "UserUpdateRoles",
+  components: { UserUpdateSfdsAuth },
   props: ['userId'],
   data() {
     return {
@@ -203,44 +108,14 @@ export default {
       selectedClient: null,
       clientRoles: [],
       effectiveClientRoles: [],
-      selectedRoles: [],
-      sfdsTableHeaders: [
-        { text: 'Mailbox', value: 'mailbox' },
-        { text: 'Uses', value: 'uses' },
-        { text: 'Permission', value: 'permission' },
-        { text: 'Actions', value: 'actions', sortable: false }
-      ],
-      sfds_authorizations: [],
-      sfdsMailboxes: ['HSCIS', 'BCMA', 'PHC', 'HOOPC'],
-      sfdsUses: ['bcma', 'hscis', 'phc', 'wda', 'hoopc', 'grp'],
-      sfdsPermissions: ['get', 'send', 'get-send', 'get-delete', 'get-send-delete'],
-      currentSfdsAuthorization: {},
-      editDialog: false,
-      deleteDialog: false,
-      editedIndex: -1,
-      defaultAuthorization: {
-        uses: [],
-        mailbox: '',
-        permission: ''
-      }
+      selectedRoles: []
     };
   },
   async created() {
     //TODO error handling
     await this.getClients();
   },
-  watch: {
-    editDialog (val) {
-      val || this.close()
-    },
-    deleteDialog (val) {
-      val || this.closeDelete()
-    },
-  },
   computed: {
-    dialogTitle () {
-      return this.editedIndex === -1 ? 'New Authorization' : 'Edit Authorization'
-    },
     itemsInColumn() {
       return Math.ceil(this.clientRoles.length / this.numberOfClientRoleColumns);
     },
@@ -249,43 +124,6 @@ export default {
     }
   },
   methods: {
-    editItem (item) {
-      this.editedIndex = this.sfds_authorizations.indexOf(item)
-      this.currentSfdsAuthorization = Object.assign({}, item)
-      this.editDialog = true
-    },
-    deleteItem: function (item) {
-      this.editedIndex = this.sfds_authorizations.indexOf(item)
-      this.currentSfdsAuthorization = Object.assign({}, item)
-      this.deleteDialog = true
-    },
-    deleteItemConfirm: function () {
-      this.sfds_authorizations.splice(this.editedIndex, 1)
-      this.closeDelete()
-    },
-    saveSfdsAuthorization: function() {
-      if (this.editedIndex > -1) {
-        Object.assign(this.sfds_authorizations[this.editedIndex], this.currentSfdsAuthorization)
-      } else {
-        this.sfds_authorizations.push(this.currentSfdsAuthorization)
-      }
-      UsersRepository.updateUser(this.$route.params.userid, {'attributes': { 'sfds_auth_1' : JSON.stringify(this.sfds_authorizations) } })
-      this.close()
-    },
-    close () {
-      this.editDialog = false
-      this.$nextTick(() => {
-        this.currentSfdsAuthorization = Object.assign({}, this.defaultAuthorization)
-        this.editedIndex = -1
-      })
-    },
-    closeDelete: function () {
-      this.deleteDialog = false
-      this.$nextTick(() => {
-        this.currentSfdsAuthorization = Object.assign({}, this.defaultAuthorization)
-        this.editedIndex = -1
-      })
-    },
     roleArrayPosition: function(col, item) {
       return (col - 1) * (this.itemsInColumn) + item - 1;
     },
@@ -389,12 +227,6 @@ export default {
 </script>
 
 <style>
-.sub-permissions {
-  margin: 20px 0px 20px 0px;
-}
-.sfds-authorizations-header {
-  margin: 22px 0px 22px 0px;
-}
 .right-gutters .col {
   padding: 10px 12px 10px 0px;
 }
